feat: hardcode mock SDK launcher
