+++ features/idcheck-wrapper/internal/build.gradle.kts
@@ -7,6 +7,7 @@ dependencies {
     implementation(libs.androidx.navigation.compose)
     implementation(libs.uk.gov.idcheck.repositories.api)
     implementation(libs.uk.gov.idcheck.ui.presentation)
+    implementation(projects.features.idcheckWrapper.internalApi)
     implementation(project(":features:resume:internal-api"))
     implementation(project(":libraries:di"))
     implementation(project(":libraries:navigation"))
